[Delete] Main Page
